+++ CHANGELOG.md
@@ -3,7 +3,7 @@ SPDX-FileCopyrightText: 2026 Stefan Majewsky <majewsky@gmx.net>
 SPDX-License-Identifier: Apache-2.0
 -->
 
-# v1.12.0 (TBD)
+# v1.12.0 (2026-07-31)
 
 Changes: